+++ python/parsing.cpp
@@ -213,7 +213,7 @@ static int64_t parse_polygons(PyObject* py_polygons, Array<Polygon*>& polygon_ar
         }
     } else if (ReferenceObject_Check(py_polygons)) {
         ((ReferenceObject*)py_polygons)
-            ->reference->polygons(true, true, -1, false, 0, polygon_array);
+            ->reference->get_polygons(true, true, -1, false, 0, polygon_array);
     } else if (PySequence_Check(py_polygons)) {
         for (int64_t i = PySequence_Length(py_polygons) - 1; i >= 0; i--) {
             PyObject* arg = PySequence_ITEM(py_polygons, i);
@@ -255,7 +255,7 @@ static int64_t parse_polygons(PyObject* py_polygons, Array<Polygon*>& polygon_ar
                 }
             } else if (ReferenceObject_Check(arg)) {
                 ((ReferenceObject*)arg)
-                    ->reference->polygons(true, true, -1, false, 0, polygon_array);
+                    ->reference->get_polygons(true, true, -1, false, 0, polygon_array);
             } else {
                 Polygon* polygon = (Polygon*)allocate_clear(sizeof(Polygon));
                 if (parse_point_sequence(arg, polygon->point_array, "") <= 0) {